pyproject.toml
[project]
name = "my-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.76.0",
    "gdsfactory>=9.32.0",
    "gdsfactory-install>=0.0.5",
    "gdspy>=1.6.13",
    "numpy>=2.4.1",
    "python-dotenv>=1.2.1",
]
